% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.R
\name{read.wofost}
\alias{read.wofost}
\title{Read WOFOST results}
\usage{
read.wofost(file)
}
\arguments{
\item{file}{The name of the file which the data are to be read from.
Valid extensions are: *.out, *.pps, or *.wps}
}
\description{
Read WOFOST output files (*.out, *.pps, or *.wps) into \R{}
}
\seealso{
\code{\link{tt2df}}
}

