% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.R
\name{tt2df}
\alias{tt2df}
\title{Fixed-format tables}
\usage{
tt2df(data, col.names, prefix, suffix)
}
\arguments{
\item{data}{Table body as vector of strings as read, e.g., by \code{readLines}}

\item{col.names}{Column names as string separated by whitespaces}

\item{prefix}{Optional data to insert before the data table}

\item{suffix}{Optional data to insert after the data table}
}
\description{
Generic function to convert a fixed-format Fortran table into a \code{data.frame}.
}
\seealso{
\code{\link{read.wofost}}
}

