% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{wofostats}
\alias{wofostats}
\alias{wofostats-package}
\title{Handling WOFOST model files}
\description{
WOFOST (WOrld FOod STudies) is a simulation model for the quantitative
analysis of the growth and production of annual field crops.
This package provides functions to handle WOFOST parameter input files and
model results using \R{}.
}
\details{
\tabular{ll}{
Version: \tab 0.1, tested with WOFOST 7.1.7\cr
Licence: \tab GPLv3\cr
}
}
\author{
Zacharias Steinmetz (\email{stei4785@uni-landau.de})
}
\references{
Boogaard HL, De Wit AJW, te Roller JA, Van Diepen CA, 2014. User's guide for
the WOFOST Control Centre 2.1 and the crop growth simulation model WOFOST 7.1.7.
Wageningen (Netherlands), Alterra, Wageningen University & Research
Centre.
}
\seealso{
\url{http://www.wageningenur.nl/en/Expertise-Services/Research-Institutes/alterra/Facilities-Products/Software-and-models/WOFOST.htm}
}

