% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write.cabo}
\alias{write.cabo}
\title{Write CABO weather files}
\usage{
write.cabo(data, country, station, stationno, lat, lon, alt, a, b,
  citation = NA, author = Sys.getenv("USERNAME"), desc = NA, file)
}
\arguments{
\item{data}{E.g. a MARS AGRI4CAST weather data.frame containing at least the
following columns: LATITUDE, LONGITUDE, ALTITUDE, DAY, TEMPERATURE_MAX,
TEMPERATURE_MIN, PRECIPITATION, RADIATION; ET0 is optional}

\item{country}{The country where the weather station is located}

\item{station}{The name of the weather station}

\item{stationno}{The number of the weather station}

\item{lat}{The latitude of the weather station}

\item{lon}{The longitude of the weather station}

\item{alt}{The altitude/elevation of the weather station}

\item{a, b}{Ångström coefficients}

\item{citation}{Reference information on the data set (optional)}

\item{author}{The author of the data set}

\item{desc}{A short description of the respective data set}

\item{file}{The path/file where the CABO output files should be written to}
}
\description{
\code{write.cabo} prints weather files as used by WOFOST e.g. from
MARS AGRI4CAST data (\url{http://agri4cast.jrc.ec.europa.eu/DataPortal/Default.aspx})
being composed of one file per year.
}

