% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read.wofost}
\alias{read.wofost}
\title{Read WOFOST results}
\usage{
read.wofost(file)
}
\arguments{
\item{file}{The name of the file which the data are to be read from.
Valid extensions are: *.out, *.pps, or *.wps}
}
\value{
Returns a \code{list} when reading detailed crop outputs (*.out), otherwise a
\code{data.frame} is generated.
}
\description{
Read WOFOST output files (*.out, *.pps, or *.wps) into \R{}
}

